Changed from two webpack configs to one
Before I used two different webpack configs depending if we were
building for development or production, and this was determined inside
the different gulp tasks.
Now I determine which build to run depending on the NODE_ENV variable
inside of the webpack config.

The loss here is that I haven’t added any way to test production on my
local machine, so that’s a todo for the future, for now I’m more sane
not having to repeat code over two files.